refactored type tests to always be imported, fixed some type errors
