+++ .eslintignore
@@ -1,6 +1,5 @@
 docs
 the_azle_book
-type_tests
 run_time_errors
 **/webpack.config.js
 dfx_generate
