+++ type_tests/candid/constructed/index.ts
@@ -1,0 +1,6 @@
+import './blob';
+import './opt';
+import './record';
+import './tuple';
+import './variant';
+import './vec';
