+++ type_tests/candid/index.ts
@@ -1,0 +1,4 @@
+import './constructed';
+import './manual';
+import './primitive';
+import './reference';
