+++ type_tests/candid/manual.ts
@@ -1,6 +1,6 @@
 import { Manual, text } from '../../src/lib/experimental';
+import { TypeMapping } from '../../src/lib/experimental/candid/type_mapping';
 import { AssertType, NotAnyAndExact, testCandidType } from '../assert_type';
-import { TypeMapping } from '../../src/lib/candid/type_mapping';
 
 testCandidType(Manual(text));
