+++ type_tests/candid/primitive/empty.ts
@@ -1,14 +1,8 @@
 import { empty } from '../../../src/lib/experimental';
-import {
-    AssertType,
-    NotAnyAndExact,
-    testCandidType,
-    testSerializable
-} from '../../assert_type';
-import { TypeMapping } from '../../../src/lib/candid/type_mapping';
+import { TypeMapping } from '../../../src/lib/experimental/candid/type_mapping';
+import { AssertType, NotAnyAndExact, testCandidType } from '../../assert_type';
 
 testCandidType(empty);
-testSerializable(empty);
 
 export type TestTypeMapping = AssertType<
     NotAnyAndExact<TypeMapping<typeof empty>, never>
