+++ type_tests/candid/primitive/floats/index.ts
@@ -1,0 +1,2 @@
+import './float32';
+import './float64';
