+++ type_tests/candid/primitive/index.ts
@@ -1,0 +1,9 @@
+import './bool';
+import './empty';
+import './floats';
+import './ints';
+import './nats';
+import './null';
+import './reserved';
+import './text';
+import './void';
