+++ type_tests/candid/primitive/ints/index.ts
@@ -1,0 +1,5 @@
+import './int';
+import './int8';
+import './int16';
+import './int32';
+import './int64';
