+++ type_tests/candid/primitive/ints/int64.ts
@@ -1,11 +1,11 @@
 import { int64 } from '../../../../src/lib/experimental';
+import { TypeMapping } from '../../../../src/lib/experimental/candid/type_mapping';
 import {
     AssertType,
     NotAnyAndExact,
     testCandidType,
     testSerializable
 } from '../../../assert_type';
-import { TypeMapping } from '../../../../src/lib/candid/type_mapping';
 
 testCandidType(int64);
 testSerializable(int64);
