+++ type_tests/candid/primitive/nats/index.ts
@@ -1,0 +1,5 @@
+import './nat';
+import './nat8';
+import './nat16';
+import './nat32';
+import './nat64';
