+++ type_tests/candid/primitive/nats/nat.ts
@@ -1,11 +1,11 @@
 import { nat } from '../../../../src/lib/experimental';
+import { TypeMapping } from '../../../../src/lib/experimental/candid/type_mapping';
 import {
     AssertType,
     NotAnyAndExact,
     testCandidType,
     testSerializable
 } from '../../../assert_type';
-import { TypeMapping } from '../../../../src/lib/candid/type_mapping';
 
 testCandidType(nat);
 testSerializable(nat);
