+++ type_tests/candid/primitive/reserved.ts
@@ -1,11 +1,11 @@
 import { reserved } from '../../../src/lib/experimental';
+import { TypeMapping } from '../../../src/lib/experimental/candid/type_mapping';
 import {
     AssertType,
     IsAny,
     testCandidType,
     testSerializable
 } from '../../assert_type';
-import { TypeMapping } from '../../../src/lib/candid/type_mapping';
 
 testCandidType(reserved);
 testSerializable(reserved);
