+++ type_tests/candid/reference/index.ts
@@ -1,0 +1,3 @@
+import './func';
+import './principal';
+import './service';
