+++ type_tests/candid/reference/principal.ts
@@ -1,11 +1,11 @@
 import { Principal } from '../../../src/lib/experimental';
+import { TypeMapping } from '../../../src/lib/experimental/candid/type_mapping';
 import {
     AssertType,
     NotAnyAndExact,
     testCandidType,
     testSerializable
 } from '../../assert_type';
-import { TypeMapping } from '../../../src/lib/candid/type_mapping';
 
 testCandidType(Principal);
 testSerializable(Principal);
