+++ type_tests/canister_methods/index.ts
@@ -1,0 +1,4 @@
+import './init';
+import './post_upgrade';
+import './query';
+import './update';
