+++ type_tests/stable_structures/index.ts
@@ -1,0 +1,1 @@
+import './stable_b_tree_map';
